builds/hackathon-tracker/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackathon-tracker"
version = "1.0.0"
description = "Discover and rank hackathon opportunities for maximum ROI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "AI SRE Portfolio", email = "portfolio@example.com"},
]
maintainers = [
    {name = "AI SRE Portfolio", email = "portfolio@example.com"},
]
keywords = ["hackathon", "opportunities", "roi", "prizes", "reddit", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
    "click>=8.0.0",
    "praw>=7.6.0",
    "python-dotenv>=0.19.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "feedparser>=6.0.0",
    "pandas>=1.3.0",
    "openpyxl>=3.0.0",
    "rich>=10.0.0",
    "schedule>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/agentbantu-wq/ai-sre-portfolio"
Repository = "https://github.com/agentbantu-wq/ai-sre-portfolio"
Issues = "https://github.com/agentbantu-wq/ai-sre-portfolio/issues"
Changelog = "https://github.com/agentbantu-wq/ai-sre-portfolio/blob/main/builds/hackathon-tracker/CHANGELOG.md"

[project.scripts]
hackathon-tracker = "hackathon_tracker.cli:cli"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["hackathon_tracker*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
hackathon_tracker = ["*.txt", "*.md"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["hackathon_tracker"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]